requirements.txt
ultralytics==8.3.2
opencv-python==4.10.0.84
torch==2.3.1
torchvision==0.18.1
pyttsx3==2.90
dspy-ai==2.5.1
numpy==1.26.4